Create Root2Tip.R
This R-code uses Root-2-tip code from APE to loop over a set of ML trees located in a folder and perform root-2-tip correlations to determine temporal signal (molecular clock).  It will out put a figure per tree that includes the phylogeny, root-2-tip regression line, and regression/correlation statistics.  It will also save a table with all of the regression statistics.  The name of the tree file will be used to label the results.